changed home and href the FAQ
